backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-environ
django-filter
Pillow==9.5.0
psycopg2-binary
whitenoise[brotli]
django-cors-headers
djoser
gunicorn